Add files via upload
This Tableau project shows the analysis for Fortune 1000 companies based on metrics like Revenue and Profits of different Sectors and Industries. 
Revenue vs Profits:
Revenue is the total amount of income generated by the sale of goods or services related to the company's primary operations.
Profit is the amount of income that remains after accounting for all expenses, debts, additional income streams, and operating costs.